Phase-III/backend/requirements.txt
openai>=1.0.0
openai-agents>=0.7.0
fastapi>=0.129.0
sqlmodel==0.0.16
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.27.0
alembic==1.13.1
pydantic-settings==2.4.0
PyJWT==2.10.1
bcrypt>=4.0.0
email-validator>=2.0.0
mcp>=1.0.0